Clone repository from https://github.com/jzamora5/AirBnB_clone_v3.git
The task requires one to fork the repository from jzamora5/AirBnB_clone_v3
and work with it as a starting point for the project. However, the repository
(jzamora5/AirBnB_clone_v3) was forked from @alexaorrico/AirBnB_clone_v2,
whose fork already exists in my account.
Given that Github prohibits forking a repository twice, I had to clone the
repository instead and work with it as a standalone repository, giving
credit to the previous authors.